requirements.txt
pywinrm==0.4.3
python-dotenv==1.0.0
paramiko==3.4.0 